Only run Ethereum Tracker end-to-end test locally
Move it from the `linera_net_tests` to the `local_net_tests`, so that it
is allowed to perform HTTP requests.
